+++ test/specs/expect/es6-nested-regex.js
@@ -1,0 +1,19 @@
+riot.tag2('es6-nested-regex', '<h1>{message}</h1>', '', '', function(opts) {
+    const x = 1
+
+    this.message = `
+      regex: ${
+        /x}`/.source
+      }`
+
+    this.foobar = `
+    foobar() {
+      return ${
+        foo( 5 /x/.1 )
+      }
+    }
+`
+
+  this.foo = function(bar) {}.bind(this)
+
+});