+++ test/specs/expect/issue_2369.js
@@ -1,8 +1,6 @@
-// riot/riot#2369
-riot.tag2('my-tag', '<h1>{getSpaceName(message)}</h1>', '', '', function(opts) {
+riot.tag2('bug-2369', '<h1>{getSpaceName(message)}</h1>', '', '', function(opts) {
     this.message = 'display/example/stuff'
-    this.getSpaceNameForLink = function(link) {
-
+    this.getSpaceName = function(link) {
         return link.match(/display\/(\w+)\//)[1]
     }
 });